fixed unpopulating edit

+++ src/components/tasks/TaskCard.js
@@ -26,18 +26,32 @@ export const TaskCard = ({ task, handleDelete, getTasks }) => {
     })
   }
 
+  if (task.userId === parseInt(sessionStorage.getItem("drifting_user"))) {
+    return (
+      <div className="card-task">
+        <div className="task-card">
+          <h2>{task.taskName}</h2>
+          <h4>{task.taskDesc}</h4>
+          {task.isCompleted ? null : <button className="tasks-delete" onClick={() => handleDelete(task.id)}>Delete</button>}
+          {task.isCompleted ? null : <button className="tasks-edit" onClick={() => history.push(`/upcoming/tasks/${task.eventId}/edit/${task.id}`)}>Edit</button>}
+        </div>
+        <div className="checkbox">
+          {task.isCompleted ? null : <input type="checkbox" name="completed" onClick={() => changeCompleted()} />}
+        </div>
+      </div>
+    )
+  } else {
 
-  return (
-    <div className="card-task">
-      <div className="task-card">
-        <h2>{task.taskName}</h2>
-        <h4>{task.taskDesc}</h4>
-        {task.isCompleted ? null : <button className="tasks-delete" onClick={() => handleDelete(task.id)}>Delete</button>}
-        {task.isCompleted ? null : <button className="tasks-edit" onClick={() => history.push(`/upcoming/tasks/${task.eventId}/edit/${task.id}`)}>Edit</button>}
+    return (
+      <div className="card-task">
+        <div className="task-card">
+          <h2>{task.taskName}</h2>
+          <h4>{task.taskDesc}</h4>
+        </div>
       </div>
-      <div className="checkbox">
-        {task.isCompleted ? null : <input type="checkbox" name="completed" onClick={() => changeCompleted()} />}
-      </div>
-    </div>
-  )
+    )
+
+  }
+
+
 }
